% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{add.coo}
\alias{add.coo}
\title{Add two coo objects element-wise}
\usage{
add.coo(x, y, a = 1, b = 1)
}
\arguments{
\item{x}{coo object}

\item{y}{coo object}

\item{a}{numeric, scalar for addition}

\item{b}{numeric, scalar for addition}
}
\value{
new coo object with (a*x) + (b*y)
}
\description{
Add two coo objects element-wise
}
