% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{center.embedding}
\alias{center.embedding}
\title{Adjust a matrix so that each column is centered around zero}
\usage{
center.embedding(x)
}
\arguments{
\item{x}{matrix}
}
\value{
matrix of same shape as x
}
\description{
Adjust a matrix so that each column is centered around zero
}
