% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{check.compatible.coo}
\alias{check.compatible.coo}
\title{Check that two coo objects are compatible for addition, multiplication}
\usage{
check.compatible.coo(x, y, msg = "")
}
\arguments{
\item{x}{object of class coo}

\item{y}{object of class coos}

\item{msg}{character, message to print alongside error}
}
\description{
Check that two coo objects are compatible for addition, multiplication
}
\keyword{internal}
