% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{check.coo}
\alias{check.coo}
\title{Check class for coo}
\usage{
check.coo(x, msg = "")
}
\arguments{
\item{x}{object of class coo}

\item{msg}{character, message to print alongside error}
}
\description{
Check class for coo
}
\keyword{internal}
