% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_learn.R
\name{check.learn.available}
\alias{check.learn.available}
\title{check whether python module is available, abort if not}
\usage{
check.learn.available()
}
\description{
check whether python module is available, abort if not
}
\keyword{internal}
