% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{clip}
\alias{clip}
\title{Force (clip) a value into a finite range}
\usage{
clip(x, xmax = 4)
}
\arguments{
\item{x}{numeric; single value or a vector}

\item{xmax}{maximum value for x}
}
\value{
numeric values in range [-xmax, xmax]
}
\description{
This implementation exists mainly to
facilitate testing the Rcpp version.
}
\keyword{internal}
