% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clip4}
\alias{clip4}
\title{perform a compound transformation on a vector, including clipping}
\usage{
clip4(x, inner, outer)
}
\arguments{
\item{x}{numeric vector}

\item{inner}{numeric constant}

\item{outer}{numeric constan}
}
\value{
numeric vector of same length as x, containing outer*clip4(inner*x)
}
\description{
perform a compound transformation on a vector, including clipping
}
\keyword{internal}
