% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{column.seeds}
\alias{column.seeds}
\title{deterministically produce random-like integers for each column in a dataset}
\usage{
column.seeds(x, key = 123)
}
\arguments{
\item{x}{matrix, items in columns}

\item{key}{numeric, a "salt" used trigger different random-like integers if needed}
}
\value{
vector of integers
}
\description{
deterministically produce random-like integers for each column in a dataset
}
\keyword{internal}
