% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_spectral.R
\name{concomp.coo}
\alias{concomp.coo}
\title{Count the number of connected components in a coo graph}
\usage{
concomp.coo(x)
}
\arguments{
\item{x}{coo object}
}
\value{
list with number of connected components and a vector
assigning rows in object to components
}
\description{
Count the number of connected components in a coo graph
}
