% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{coo}
\alias{coo}
\title{Create a coo representation of a square matrix}
\usage{
coo(x)
}
\arguments{
\item{x}{square matrix}
}
\value{
matrix with three columns (from index, to index, value)
}
\description{
Create a coo representation of a square matrix
}
\keyword{internal}
