% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{coo2mat}
\alias{coo2mat}
\title{Convert from coo object into conventional matrix}
\usage{
coo2mat(x)
}
\arguments{
\item{x}{coo object}
}
\value{
matrix
}
\description{
Convert from coo object into conventional matrix
}
