% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dCenteredPearson}
\alias{dCenteredPearson}
\title{compute pearson correlation distance between two vectors}
\usage{
dCenteredPearson(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
pearson distance between x and y
}
\description{
Pearson distance is (1-r^2)
}
\details{
Important: this function assumes that data has been centered
i.e. that mean(x) = mean(y) = 0
}
\keyword{internal}
