% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dCosine}
\alias{dCosine}
\title{compute cosine dissimilarity between two vectors}
\usage{
dCosine(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
cosine dissimilarity between x and y
}
\description{
Note: values output from this function do not satisfy the triangle inequality
}
\keyword{internal}
