% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_learn.R
\name{detect.umap.learn}
\alias{detect.umap.learn}
\title{adjust config depending on umap-learn version}
\usage{
detect.umap.learn(config)
}
\arguments{
\item{config}{list with settings}
}
\value{
config list with set umap_learn_args
}
\description{
adjust config depending on umap-learn version
}
\keyword{internal}
