% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dEuc2}
\alias{dEuc2}
\title{compute squared Euclidean distance between two vectors}
\usage{
dEuc2(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
square of euclidean distance
}
\description{
compute squared Euclidean distance between two vectors
}
