% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dEuclidean}
\alias{dEuclidean}
\title{compute Euclidean distance between two vectors}
\usage{
dEuclidean(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
euclidean norm of x-y
}
\description{
compute Euclidean distance between two vectors
}
