% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dManhattan}
\alias{dManhattan}
\title{compute Manhattan distance between two vectors}
\usage{
dManhattan(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
manhattan norm of x-y
}
\description{
compute Manhattan distance between two vectors
}
