% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{eucd}
\alias{eucd}
\title{Compute euclidean distances between two vectors (in any dimension)}
\usage{
eucd(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
numeric, euclidean distance
}
\description{
(For performance reasons, this function does not perform any checks)
(Tried Rcpp here, but does not provide much speedup)
}
