% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{eucd2}
\alias{eucd2}
\title{Compute squared distance between two vectors}
\usage{
eucd2(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
numeric, squared distance
}
\description{
(For performance reasons, this function does not perform any checks)
(Tried Rcpp here, but does not provide much speedup)
}
