% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{eucmag2}
\alias{eucmag2}
\title{Compute squared euclidean magnitude of a vector}
\usage{
eucmag2(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric, squared magnitude
}
\description{
Compute squared euclidean magnitude of a vector
}
