% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{find.ab.params}
\alias{find.ab.params}
\title{Estimate a/b parameters}
\usage{
find.ab.params(spread, min.dist, alim = c(0, 20), blim = c(0, 20),
  tolerance = 1e-08)
}
\arguments{
\item{spread}{numeric}

\item{min.dist}{numeric}

\item{alim}{numeric vector of length 2, initial search range for parameter a}

\item{blim}{numeric vector of length 2, initial search range for parameter b}

\item{tolerance}{numeric, determines how deeply to search}
}
\value{
vector with componets "a" and "b" that are most appropriate
for the given spread and min.dist.
}
\description{
This performs a brute-force search in parameter space.
The algorithm assumes a narrowing-down search will produce a decent fit.
}
\keyword{internal}
