% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_naive.R
\name{fuzzy.simplicial.set}
\alias{fuzzy.simplicial.set}
\title{create a simplicial set from a distance object}
\usage{
fuzzy.simplicial.set(d, config = umap.defaults)
}
\arguments{
\item{d}{distance object}

\item{config}{list with settings}
}
\description{
create a simplicial set from a distance object
}
