% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_spectral.R
\name{identity.coo}
\alias{identity.coo}
\title{Construct an identity matrix}
\usage{
identity.coo(n.elements, names = NULL)
}
\arguments{
\item{n.elements}{integer, number of elements}

\item{names}{character vector, names associated with the elements}
}
\value{
new coo object
}
\description{
Construct an identity matrix
}
