% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn.from.data.reps}
\alias{knn.from.data.reps}
\title{Repeat knn.from.data multiple times, pick the best neighbors}
\usage{
knn.from.data.reps(d, k, metric.function, subsample.k = 0.5, reps = 2)
}
\arguments{
\item{d}{matrix with data}

\item{k}{integer, number of neighbors}

\item{metric.function}{function with signature f(a, b) that returns a metric distance}

\item{subsample.k}{numeric, used for internal tuning of implementation}

\item{reps}{integer, number of repeats to carry out}
}
\value{
list of same format as knn.from.data
}
\description{
Repeat knn.from.data multiple times, pick the best neighbors
}
\keyword{internal}
