% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn.from.dist}
\alias{knn.from.dist}
\title{get information about k nearest neighbors from a distance object or from a matrix
with distances}
\usage{
knn.from.dist(d, k)
}
\arguments{
\item{d}{dist object or matrix with distances}

\item{k}{integer, number of neighbors}
}
\value{
list with two components;
indexes identifies, for each point in dataset, the set of k neighbors
distances provides distances from each point to those neighbors
}
\description{
This implementation uses sorting of distances to identify the k elements 
that are nearest to each data point. The result is deterministic and exact.
}
\details{
By definition, the first nearest neighbor to each point is the point itself.
Subsequent neighbors are "true" neighbors.
}
