% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{knn.info}
\alias{knn.info}
\title{Compute knn information}
\usage{
knn.info(d, config)
}
\arguments{
\item{d}{data matrix}

\item{config}{list with settings; relevant settings are as follows:
input - "data" or "dist"
n.neighbors - number of neighbors k
metric.function - function with signature f(a, b)}
}
\value{
list with at least two components, indexes and distances
}
\description{
This function determines whether to obtain knn information using an exact
brute force approach or using an approximate algorithm
}
\keyword{internal}
