% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_spectral.R
\name{lanczos.coo}
\alias{lanczos.coo}
\title{Construct matrix with k Lanczos vectors from a real Hermitian matrix x}
\usage{
lanczos.coo(x, k)
}
\arguments{
\item{x}{coo object}

\item{k}{number of vectors}
}
\description{
Construct matrix with k Lanczos vectors from a real Hermitian matrix x
}
