% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_spectral.R
\name{laplacian.coo}
\alias{laplacian.coo}
\title{Construct a normalized Laplacian for a graph}
\usage{
laplacian.coo(x)
}
\arguments{
\item{x}{coo object encoding a graph}
}
\value{
new coo object
}
\description{
This implementation constructs the laplacian  element-by-element.
Diagonals: 1, Element_ij = -1/sqrt(deg_i deg_j)
}
