% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{make.coo}
\alias{make.coo}
\title{Helper to construct coo objects}
\usage{
make.coo(x, names, n.elements)
}
\arguments{
\item{x}{coo matrix}

\item{names}{character vector}

\item{n.elements}{integer}
}
\value{
coo object
}
\description{
Helper to construct coo objects
}
\keyword{internal}
