% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{make.epochs.per.sample}
\alias{make.epochs.per.sample}
\title{Compute a value to capture how often each item contributes to layout optimization}
\usage{
make.epochs.per.sample(w, epochs)
}
\arguments{
\item{w}{numeric vector or matrix}

\item{epochs}{integer}
}
\value{
numeric vector of same length as w
}
\description{
Compute a value to capture how often each item contributes to layout optimization
}
\keyword{internal}
