% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{make.initial.embedding}
\alias{make.initial.embedding}
\title{Create an initial embedding for a graph}
\usage{
make.initial.embedding(V, config)
}
\arguments{
\item{V}{integer, number of vertices}

\item{config}{list with settings}
}
\value{
matrix with an embedding
}
\description{
This either takes a set embedding from config, or sets a random state
}
