% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{make.initial.spectator.embedding}
\alias{make.initial.spectator.embedding}
\title{Create an initial embedding for a set of spectators}
\usage{
make.initial.spectator.embedding(embedding, knn.indexes)
}
\arguments{
\item{embedding}{matrix with an existing (primary) embedding}

\item{knn.indexes}{matrix with indexes from spectators to components primary embedding}
}
\value{
matrix with an embedding
}
\description{
Create an initial embedding for a set of spectators
}
\keyword{internal}
