% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{make.spectral.embedding}
\alias{make.spectral.embedding}
\title{Create a spectral embedding for a connectivity graph}
\usage{
make.spectral.embedding(V, d, g)
}
\arguments{
\item{V}{integer, number of vertices in embedding}

\item{d}{integer, number of dimensions}

\item{g}{coo object}
}
\value{
embedding matrix
}
\description{
Create a spectral embedding for a connectivity graph
}
