% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdCenteredPearson}
\alias{mdCenteredPearson}
\title{compute pearson correlation distances}
\usage{
mdCenteredPearson(m, origin, targets)
}
\arguments{
\item{m}{matrix with raw data}

\item{origin}{index (1-based) of origin element}

\item{targets}{indexes (1-based) of target elements}
}
\value{
dCenteredPearson norms between origin and targets
}
\description{
compute pearson correlation distances
}
\keyword{internal}
