% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdCosine}
\alias{mdCosine}
\title{compute cosine distances}
\usage{
mdCosine(m, origin, targets)
}
\arguments{
\item{m}{matrix with raw data}

\item{origin}{index (1-based) of origin element}

\item{targets}{indexes (1-based) of target elements}
}
\value{
dCosine norms between origin and targets
}
\description{
compute cosine distances
}
\keyword{internal}
