% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdEuclidean}
\alias{mdEuclidean}
\title{compute Euclidean distances between one origin and several targets}
\usage{
mdEuclidean(m, origin, targets)
}
\arguments{
\item{m}{matrix with raw data}

\item{origin}{integer index of origin (1-based)}

\item{targets}{array of indexes to targets (1-based)}
}
\value{
dEuclidean norms between origin and targets
}
\description{
compute Euclidean distances between one origin and several targets
}
