% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdManhattan}
\alias{mdManhattan}
\title{compute Manhattan distances between one origin and several targets}
\usage{
mdManhattan(m, origin, targets)
}
\arguments{
\item{m}{matrix with raw data}

\item{origin}{integer index of origin (1-based)}

\item{targets}{array of indexes to targets (1-based)}
}
\value{
dManhattan norms between origin and targets
}
\description{
compute Manhattan distances between one origin and several targets
}
