% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{multiplicationprep.coo}
\alias{multiplicationprep.coo}
\title{Prepare a coo object by splitting a coo}
\usage{
multiplicationprep.coo(x)
}
\arguments{
\item{x}{coo object}
}
\value{
list with two components $to and $from
}
\description{
Prepare a coo object by splitting a coo
}
\keyword{internal}
