% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_naive.R
\name{naive.fuzzy.simplicial.set}
\alias{naive.fuzzy.simplicial.set}
\title{create a simplicial set from a distance object}
\usage{
naive.fuzzy.simplicial.set(knn, config)
}
\arguments{
\item{knn}{list with inform about nearest neighbors (output of knn.info)}

\item{config}{list with settings}
}
\value{
matrix
}
\description{
create a simplicial set from a distance object
}
\keyword{internal}
