% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_naive.R
\name{naive.optimize.embedding}
\alias{naive.optimize.embedding}
\title{modify an existing embedding}
\usage{
naive.optimize.embedding(tembedding, config, eps)
}
\arguments{
\item{tembedding}{matrix, transpose of matrix with an initial embedding}

\item{config}{list with settings}

\item{eps}{matrix with connectivity coo graph and epochs per sample;
this must be a matrix with columns "from", "to", "value", and "eps"}
}
\value{
matrix of same dimension as initial tembedding
}
\description{
modify an existing embedding
}
\keyword{internal}
