% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_naive.R
\name{naive.simplicial.set.embedding}
\alias{naive.simplicial.set.embedding}
\title{create an embedding of graph into a low-dimensional space}
\usage{
naive.simplicial.set.embedding(g, config)
}
\arguments{
\item{g}{matrix, graph connectivity as coo}

\item{config}{list with settings}
}
\value{
matrix with embedding,
nrows is from g, ncols determined from config
}
\description{
create an embedding of graph into a low-dimensional space
}
