% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimize_embedding}
\alias{optimize_embedding}
\title{run a series of epochs of the umap optimization}
\usage{
optimize_embedding(embedding, pairs, eps, epns, abg, alpha0, num_epochs)
}
\arguments{
\item{embedding}{matrix with embedding coordinates 
(This should be transpose of final embedding, with coordinates along columns, items in rows)}

\item{pairs}{matrix with two columns; rows should linke to (from, to); indexes 0-based}

\item{eps}{numeric vector, epochs for next sample}

\item{epns}{numeric vector, epochs for next negative sample}

\item{abg}{vector with configuration parameters, a, b, gamma, move_other}

\item{alpha0}{numeric, initial learning rate}
}
\description{
run a series of epochs of the umap optimization
}
\keyword{internal}
