% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimize_epoch}
\alias{optimize_epoch}
\title{run one epoch of the umap optimization}
\usage{
optimize_epoch(embedding, pairs, adjust, nns, abg, alpha)
}
\arguments{
\item{embedding}{matrix with embedding coordinates
(This should be transpose of final embedding, with coordinates along columns, items in rows)}

\item{pairs}{matrix with two columns; rows should link to (from to); indexes 0-based}

\item{adjust}{vector with 0/1 whether to adjust or not}

\item{nns}{vector with negative-neighbors-set size}

\item{abg}{vector with configuration parameters, a, b, gamma, move_other}

\item{alpha}{numeric learning rate for this epoch}
}
\description{
run one epoch of the umap optimization
}
\keyword{internal}
