% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_print.R
\name{print.umap.knn}
\alias{print.umap.knn}
\title{Display summary of knn.info}
\usage{
\method{print}{umap.knn}(x, ...)
}
\arguments{
\item{x}{object of class umap.knn}

\item{...}{ignored}
}
\description{
Display summary of knn.info
}
\keyword{internal}
