% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_print.R
\name{print.umap}
\alias{print.umap}
\title{Display a summary of a umap object}
\usage{
\method{print}{umap}(x, ...)
}
\arguments{
\item{x}{umap object}

\item{...}{other parameters (not used)}
}
\description{
Display a summary of a umap object
}
\keyword{internal}
