% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{reduce.coo}
\alias{reduce.coo}
\title{Remove some entires in a coo matrix where values are zero}
\usage{
reduce.coo(x)
}
\arguments{
\item{x}{coo object}
}
\value{
matrix based on x, perhaps with some lines in original removed
}
\description{
Remove some entires in a coo matrix where values are zero
}
\keyword{internal}
