% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.global.seed}
\alias{set.global.seed}
\title{set .Random.seed to a pre-saved value}
\usage{
set.global.seed(x)
}
\arguments{
\item{x}{integer vector}
}
\description{
set .Random.seed to a pre-saved value
}
\keyword{internal}
