% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_naive.R
\name{smooth.knn.dist}
\alias{smooth.knn.dist}
\title{compute a "smooth" distance to the kth neighbor and approximate first neighbor}
\usage{
smooth.knn.dist(k.dist, neighbors, iterations = 64,
  local.connectivity = 1, bandwidth = 1, tolerance = 1e-05,
  min.dist.scale = 0.001)
}
\arguments{
\item{k.dist}{matrix with distances to k neighbors}

\item{neighbors}{numeric, number of neighbors to approximate for}

\item{iterations}{integers, number of iterations}

\item{local.connectivity}{iteger}

\item{bandwidth}{numeric}

\item{tolerance}{numeric, for numeric precision}

\item{min.dist.scale}{numeric, minimum distance to nearest neighbor (for display)}
}
\value{
list with two vectors, distances to the kth neighbor,
and distance to the first nearest neighbor
}
\description{
compute a "smooth" distance to the kth neighbor and approximate first neighbor
}
\keyword{internal}
