% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{spectator.knn.info}
\alias{spectator.knn.info}
\title{compute knn information for spectators relative to data}
\usage{
spectator.knn.info(spectators, d, config, brute.force = TRUE)
}
\arguments{
\item{spectators}{matrix with data for spectators}

\item{d}{matrix with primary data}

\item{config}{list with settings}

\item{brute.force}{logical, during developemnt, set FALSE to force stochastic knn}
}
\value{
list with at least two components, indexes and distances
}
\description{
compute knn information for spectators relative to data
}
\keyword{internal}
