% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_spectral.R
\name{spectral.coo}
\alias{spectral.coo}
\title{Get a set of k eigenvalues and eigenvectors}
\usage{
spectral.coo(x, k, m = 2 * k + 1)
}
\arguments{
\item{x}{coo object}

\item{k}{integer, number of eigenvalues/eigenvectors}

\item{m}{integer, number of lanczos vectors to use
(The higher the number the better, 2k+1 should be the minimum)}
}
\value{
list with two components
}
\description{
Get a set of k eigenvalues and eigenvectors
}
\keyword{internal}
