% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{spectral.eigenvectors}
\alias{spectral.eigenvectors}
\title{get a set of k eigenvectors for the laplacian of x}
\usage{
spectral.eigenvectors(x, k)
}
\arguments{
\item{x}{coo object}

\item{k}{integer}
}
\value{
list with
}
\description{
get a set of k eigenvectors for the laplacian of x
}
\keyword{internal}
