% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{stop.coo}
\alias{stop.coo}
\title{Stop execution with a custom message}
\usage{
stop.coo(msg1, msg2 = "")
}
\arguments{
\item{msg1}{character}

\item{msg2}{character}
}
\description{
Stop execution with a custom message
}
