% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_spectral.R
\name{subset.coo}
\alias{subset.coo}
\title{Subset a coo}
\usage{
\method{subset}{coo}(x, items)
}
\arguments{
\item{x}{coo object}

\item{items}{items (indexes) to keep}
}
\value{
new smaller coo object
}
\description{
Subset a coo
}
