% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{t.coo}
\alias{t.coo}
\title{Transpose a coo matrix}
\usage{
\method{t}{coo}(x)
}
\arguments{
\item{x}{coo object}
}
\value{
another coo object describing a transposed matrix
}
\description{
Transpose a coo matrix
}
\keyword{internal}
