% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_checks.R
\name{umap.check.config.class}
\alias{umap.check.config.class}
\title{Validator for config class component}
\usage{
umap.check.config.class(config)
}
\arguments{
\item{config}{list with arguments (object of class umap.config)}
}
\description{
Validator for config class component
}
\keyword{internal}
