% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_checks.R
\name{umap.check.input}
\alias{umap.check.input}
\title{Validator for umap primary input}
\usage{
umap.check.input(d)
}
\arguments{
\item{d}{matrix or compatible}
}
\value{
d as matrix
}
\description{
Validator for umap primary input
}
