% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_cmdscale.R
\name{umap.cmdscale}
\alias{umap.cmdscale}
\title{create an embedding using cmdscale}
\usage{
umap.cmdscale(d, config)
}
\arguments{
\item{d}{distance object}

\item{config}{list with settings; ignored in this implementation}
}
\description{
create an embedding using cmdscale
}
