% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\docType{data}
\name{umap.defaults}
\alias{umap.defaults}
\title{Default configuration for umap}
\format{An object of class \code{umap.config} of length 18.}
\usage{
umap.defaults
}
\arguments{
\item{n.neighbors}{integer}

\item{n.components}{integer}
}
\description{
Default
}
\keyword{datasets}
