% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_naive.R
\name{umap.naive.predict}
\alias{umap.naive.predict}
\title{predict embedding of new data given an existing umap object}
\usage{
umap.naive.predict(umap, data)
}
\arguments{
\item{umap}{object of class umap}

\item{data}{matrix with new data}
}
\value{
matrix with embedding coordinates
}
\description{
predict embedding of new data given an existing umap object
}
