% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_naive.R
\name{umap.naive}
\alias{umap.naive}
\title{Create a umap embedding}
\usage{
umap.naive(d, config)
}
\arguments{
\item{d}{data object}

\item{config}{list with settings}
}
\value{
list, one element of which is matrix with embedding coordinates
}
\description{
This implementation is called naive because it is a rather straightforward
translation of the original python code.
}
