% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_checks.R
\name{umap.prep.input}
\alias{umap.prep.input}
\title{Prep primary input as a data matrix}
\usage{
umap.prep.input(d, config)
}
\arguments{
\item{d}{matrix or compatible}

\item{config}{list with settings}
}
\value{
d as matrix
}
\description{
Prep primary input as a data matrix
}
\keyword{internal}
