% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_python.R
\name{umap.python}
\alias{umap.python}
\title{Create a umap embedding using a python package}
\usage{
umap.python(d, config)
}
\arguments{
\item{d}{data object}

\item{config}{list with settings}
}
\value{
list, one element of which is matrix with embedding coordinates
}
\description{
Create a umap embedding using a python package
}
\keyword{internal}
