% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{umap}
\alias{umap}
\title{compute a manifold approximation and projection}
\usage{
umap(d, config = umap.defaults, method = c("naive", "cmdscale"), ...)
}
\arguments{
\item{d}{matrix}

\item{config}{object of class umap.config}

\item{method}{character, implementation}

\item{...}{list of settings; overwrite settings in config}
}
\description{
compute a manifold approximation and projection
}
