% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_small.R
\name{umap.small}
\alias{umap.small}
\title{Create an embedding object compatible with package umap for very small inputs}
\usage{
umap.small(d, config)
}
\arguments{
\item{d}{matrix}

\item{config}{list with settings}
}
\value{
list, one element of which is matrix with embedding coordinates
}
\description{
Create an embedding object compatible with package umap for very small inputs
}
\keyword{internal}
