% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_checks.R
\name{umap.warning}
\alias{umap.warning}
\title{create a warning message}
\usage{
umap.warning(...)
}
\arguments{
\item{...}{strings for error message}
}
\description{
create a warning message
}
\keyword{internal}
