% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_universal.R
\name{vector.norm}
\alias{vector.norm}
\title{Compute vector norm}
\usage{
vector.norm(z)
}
\arguments{
\item{z}{numeric vector}
}
\value{
numeric, vector norm
}
\description{
Compute vector norm
}
\keyword{internal}
