% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_naive.R
\name{knn.info}
\alias{knn.info}
\title{get information about k nearest neighbors}
\usage{
knn.info(dm, neighbors)
}
\arguments{
\item{dm}{matrix with distances}

\item{neighbors}{integer, number of neighbors}
}
\value{
list with neighbor indeces and distances
}
\description{
get information about k nearest neighbors
}
