% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_spectral.R
\name{lanczos.coo}
\alias{lanczos.coo}
\title{Construct matrix with k Lanczos vectors from a real Hermitian matrix x}
\usage{
lanczos.coo(x, k, m)
}
\arguments{
\item{x}{coo object}

\item{k}{integer, number of vectors to optimize}

\item{m}{integer, number of vectors to use in procedure (set higher than k)}
}
\value{
list with vectors and transfer matrix
}
\description{
Based on description on wikipedia:
https://en.wikipedia.org/wiki/Lanczos_algorithm
with restart implemented ad-hoc
}
\details{
This does not work very well (eigenvectors don't match svd()). Help would be appreciated.
}
\keyword{internal}
