% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{make.initial.embedding}
\alias{make.initial.embedding}
\title{Create an initial embedding for a graph}
\usage{
make.initial.embedding(V, config, g = NULL)
}
\arguments{
\item{V}{integer, number of vertices}

\item{config}{list with settings}

\item{g}{coo object with graph connectivity}
}
\value{
matrix with an embedding
}
\description{
This either takes a set embedding from config, or sets a random state
}
\keyword{internal}
