% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{make.spectral.embedding}
\alias{make.spectral.embedding}
\title{Create a spectral embedding for a connectivity graph}
\usage{
make.spectral.embedding(d, g)
}
\arguments{
\item{d}{integer, number of dimensions}

\item{g}{coo object}
}
\value{
embedding matrix. Might return NULL if spectral embedding fails
}
\description{
Create a spectral embedding for a connectivity graph
}
\keyword{internal}
