% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdCenteredPearson}
\alias{mdCenteredPearson}
\title{compute pearson correlation distances}
\usage{
mdCenteredPearson(m)
}
\arguments{
\item{m}{matrix with raw data}
}
\value{
dCenteredPearson norms between first row and all other rows
}
\description{
compute pearson correlation distances
}
