% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdCosine}
\alias{mdCosine}
\title{compute cosine distances}
\usage{
mdCosine(m)
}
\arguments{
\item{m}{matrix with raw data}
}
\value{
dCosine norms between first row and all other rows
}
\description{
compute cosine distances
}
\keyword{internal}
