% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdCosine}
\alias{mdCosine}
\title{compute cosine distances between one origin and several targets}
\usage{
mdCosine(m, origin, targets)
}
\arguments{
\item{m}{matrix with raw data}

\item{origin}{integer index of origin (1-based)}

\item{targets}{array of indexes to targets (1-based)}
}
\value{
dCosine norms between origin and targets
}
\description{
compute cosine distances between one origin and several targets
}
