% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdEuclidean}
\alias{mdEuclidean}
\title{compute Euclidean distances}
\usage{
mdEuclidean(m, origin, targets)
}
\arguments{
\item{m}{matrix with raw data}

\item{origin}{index (1-based) of origin element}

\item{targets}{indexes (1-based) of target elements}
}
\value{
dEuclidean norms between origin and target elements
}
\description{
compute Euclidean distances
}
\keyword{internal}
