% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdEuclidean}
\alias{mdEuclidean}
\title{compute Euclidean distances}
\usage{
mdEuclidean(m)
}
\arguments{
\item{m}{matrix with raw data}
}
\value{
dEuclidean norms between first row and all other rows
}
\description{
compute Euclidean distances
}
\keyword{internal}
