% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdManhattan}
\alias{mdManhattan}
\title{compute Manhattan distances}
\usage{
mdManhattan(m, origin, targets)
}
\arguments{
\item{m}{matrix with raw data}

\item{origin}{index (1-based) of origin element}

\item{targets}{indexes (1-based) of target elements}
}
\value{
dManhattan norms between origin and targets
}
\description{
compute Manhattan distances
}
\keyword{internal}
