% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdManhattan}
\alias{mdManhattan}
\title{compute Manhattan distances}
\usage{
mdManhattan(m)
}
\arguments{
\item{m}{matrix with raw data}
}
\value{
dManhattan norms between origin and targets
}
\description{
compute Manhattan distances
}
\keyword{internal}
