% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimize_epoch}
\alias{optimize_epoch}
\title{run one epoch of the umap optimization}
\usage{
optimize_epoch(embedding, pairs, adjust, nns, abg, alpha)
}
\arguments{
\item{embedding}{matrix with embedding coordinates (coordinates along columns)}

\item{pairs}{matrix with rows: from to; indexes must be 0-based}

\item{adjust}{vector with 0/1 whether to adjust or not}

\item{nns}{vector with negative-neighbors-set size}

\item{abg}{vector with configuration parameters, a, b, gamma}

\item{alpha}{numeric learning rate for this epoch}

\item{book}{matrix with columns: eps, epns, eon2s, eons, nns (in that order)}
}
\description{
run one epoch of the umap optimization
}
\keyword{internal}
