% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_print.R
\name{print.umap.config}
\alias{print.umap.config}
\title{Display contents of a umap configuration}
\usage{
\method{print}{umap.config}(x, ...)
}
\arguments{
\item{x}{object of class umap.config}

\item{...}{ignored}
}
\description{
Display contents of a umap configuration
}
\keyword{internal}
