% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo_spectral.R
\name{spectral.coo}
\alias{spectral.coo}
\title{Get a set of k eigenvalues and eigenvectors}
\usage{
spectral.coo(x, k)
}
\arguments{
\item{x}{coo object}

\item{k}{integer, number of eigenvalues/eigenvectors}
}
\value{
list with two components
}
\description{
Get a set of k eigenvalues and eigenvectors
}
