% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_checks.R
\name{umap.check.config}
\alias{umap.check.config}
\title{Validator functions for umap settings}
\usage{
umap.check.config(config = umap.defaults, ...)
}
\arguments{
\item{config}{list with umap arguments}

\item{...}{other arguments}
}
\value{
config object, may contain some different components from config in input
}
\description{
Validator functions for umap settings
}
\keyword{internal}
