% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\docType{data}
\name{umap.defaults}
\alias{umap.defaults}
\title{default configuration for umap}
\format{An object of class \code{umap.config} of length 11.}
\usage{
umap.defaults
}
\description{
default configuration for umap
}
\keyword{datasets}
