% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_checks.R
\name{umap.error}
\alias{umap.error}
\title{stop execution with a custom error message}
\usage{
umap.error(...)
}
\arguments{
\item{...}{strings for error message}
}
\description{
stop execution with a custom error message
}
