% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{umap}
\alias{umap}
\title{Computes a manifold approximation and projection}
\usage{
umap(d, config = umap.defaults, method = c("naive", "python"), ...)
}
\arguments{
\item{d}{matrix, input data}

\item{config}{object of class umap.config}

\item{method}{character, implementation. Available methods are 'naive'
(an implementation written in pure R) and 'python' (requires python package 'umap')}

\item{...}{list of settings; overwrite settings in config}
}
\value{
object of class umap, containing at least a component
with an embedding and a component with configuration settings
}
\description{
Computes a manifold approximation and projection
}
