% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coo.R
\name{vectormultiplication.coo}
\alias{vectormultiplication.coo}
\title{Matrix multiplication of a coo matrix with a vector}
\usage{
vectormultiplication.coo(x, v)
}
\arguments{
\item{x}{coo object}

\item{v}{numeric vector}
}
\value{
new vector x*v
}
\description{
Matrix multiplication of a coo matrix with a vector
}
